Level created + see description for added details
Scene I created them in doesn't have a player or a good camera setup as it is a blocking. The interactuables, can't be interacted with but are inside the level so we can know where they should be located. TO ADD: Wall blocks so the player can't jjump outside the level